♻️ Refactor: 에러 상수화 수정
